Fix clear bug and generated code

+++ generator/internal/template.go
@@ -79,13 +79,13 @@ func (s {{$ttName}}Buffer) SubSlice(low int, high int) {{$ttName}}Buffer {
 	var tVar {{$ttName}}
 	tSize := unsafe.Sizeof(tVar)
 	type internalPtr struct{
-		offset    uint32
+		offset    uintptr
 		bucketIdx uint8
 		arenaMask uint16
 	}
 	currentPtr := *(*internalPtr)(unsafe.Pointer(&s.data))
 	newPtr := internalPtr{
-			offset:    currentPtr.offset + uint32(low*int(tSize)),
+			offset:    currentPtr.offset + uintptr(low*int(tSize)),
 			bucketIdx: currentPtr.bucketIdx,
 			arenaMask: currentPtr.arenaMask,
 	}
@@ -109,13 +109,13 @@ func (s {{$ttName}}Buffer) Get(idx int) {{$ttName}}Ptr {
 	var tVar {{$ttName}}
 	tSize := unsafe.Sizeof(tVar)
 	type internalPtr struct{
-		offset    uint32
+		offset    uintptr
 		bucketIdx uint8
 		arenaMask uint16
 	}
 	currentPtr := *(*internalPtr)(unsafe.Pointer(&s.data))
 	newPtr := internalPtr{
-			offset:    currentPtr.offset + uint32(idx*int(tSize)),
+			offset:    currentPtr.offset + uintptr(idx*int(tSize)),
 			bucketIdx: currentPtr.bucketIdx,
 			arenaMask: currentPtr.arenaMask,
 	}
